# api_gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.AddToFavoritesResponse:
    properties:
      success:
        type: boolean
    type: object
...
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.CreatePostRequest:
    properties:
      content:
        minLength: 1
        type: string
      status:
        enum:
        - draft
        - published
        - archived
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - content
    - title
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.CreatePostResponse:
    properties:
      post:
        $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.Post'
    type: object
...
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.DeletePostResponse:
    properties:
      success:
        type: boolean
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.GetPostResponse:
    properties:
      post:
        $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.Post'
    type: object
...
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ListPostsResponse:
    properties:
      next_page_token:
        type: string
      posts:
        items:
          $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.Post'
        type: array
      total_count:
        type: integer
    type: object
...
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.RatePostResponse:
    properties:
      post:
        $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.Post'
    type: object
...
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.UpdatePostRequest:
    properties:
      post:
        $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.PostUpdate'
    required:
    - post
    type: object
  github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.UpdatePostResponse:
    properties:
      post:
        $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.Post'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Gateway для Mentor Platform
  title: Mentor Platform API
  version: "1.0"
paths:
...
  /posts:
    get:
      description: Возвращает список постов с фильтрацией и пагинацией
      parameters:
      - description: Токен пагинации
        in: query
        name: page_token
        type: string
      - default: 20
        description: Размер страницы
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: ID автора
        in: query
        name: author_id
        type: string
      - description: Статус поста
        enum:
        - draft
        - published
        - archived
        in: query
        name: status
        type: string
      - collectionFormat: csv
        description: Теги
        in: query
        items:
          type: string
        name: tags
        type: array
      - description: Поисковый запрос
        in: query
        name: search
        type: string
      - default: created_at
        description: Поле сортировки
        enum:
        - created_at
        - updated_at
        - title
        in: query
        name: sort_field
        type: string
      - default: desc
        description: Порядок сортировки
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ListPostsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
      summary: Получить список постов
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Создает новый пост
      parameters:
      - description: Данные для создания поста
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.CreatePostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать пост
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Удаляет пост по его идентификатору
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.DeletePostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить пост
      tags:
      - posts
    get:
      description: Возвращает информацию о посте по его идентификатору
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.GetPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
      summary: Получить пост по ID
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Обновляет информацию о посте
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.UpdatePostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить пост
      tags:
      - posts
  /posts/{id}/favorite:
    post:
      consumes:
      - application/json
      description: Добавляет пост в список избранных постов пользователя. Можно добавлять
        только опубликованные посты.
      parameters:
      - description: ID поста
        example: '"550e8400-e29b-41d4-a716-446655440000"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.AddToFavoritesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "404":
          description: Пост не найден
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "409":
          description: Пост уже в избранном
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "412":
          description: Пост не опубликован
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Добавить пост в избранное
      tags:
      - posts
  /posts/{id}/rate:
    post:
      consumes:
      - application/json
      description: Добавляет оценку к посту
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      - description: Данные для оценки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.RatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.RatePostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Оценить пост
      tags:
      - posts
  /posts/{id}/ratings:
    get:
      description: Возвращает список отзывов и рейтингов для поста с пагинацией
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      - description: Токен пагинации
        in: query
        name: page_token
        type: string
      - default: 20
        description: Размер страницы
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.GetPostRatingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_hohlayder_Mentor_Platform_api_gateway_internal_domain.ErrorResponse'
      summary: Получить отзывы поста
      tags:
      - posts
...
securityDefinitions:
  BearerAuth:
    description: 'Введите: Bearer {jwt_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"